+++ src/vue/modules/create-asset-form/components/advanced-step-form.vue
@@ -4,93 +4,6 @@
     @submit.prevent="isFormValid() && setConfirmationState()"
   >
     <!-- eslint-disable-next-line max-len -->
-    <h3 class="advanced-step-form__subheading app__form-subheading">
-      {{ 'create-asset-form.stellar-integration-subheading' | globalize }}
-    </h3>
-
-    <div class="app__form-row">
-      <div class="app__form-field">
-        <!-- stellar integration is fully coming in 1.9.0 -->
-        <tick-field
-          class="advanced-step-form__stellar-integration-tick-field"
-          v-model="form.isStellarIntegrationEnabled"
-          :disabled="isDisabled"
-          :cb-value="true"
-        >
-          {{ 'create-asset-form.integration-with-stellar-check' | globalize }}
-        </tick-field>
-      </div>
-    </div>
-
-    <template v-if="form.isStellarIntegrationEnabled">
-      <div class="app__form-row">
-        <div class="app__form-field">
-          <tick-field
-            v-model="form.stellar.deposit"
-            :disabled="isDisabled"
-            :cb-value="true"
-          >
-            {{ 'create-asset-form.deposit-lbl' | globalize }}
-          </tick-field>
-        </div>
-      </div>
-
-      <div class="app__form-row">
-        <div class="app__form-field">
-          <tick-field
-            v-model="form.stellar.withdraw"
-            :disabled="isDisabled"
-            :cb-value="true"
-          >
-            {{ 'create-asset-form.withdraw-lbl' | globalize }}
-          </tick-field>
-        </div>
-      </div>
-
-      <div class="app__form-row">
-        <div class="app__form-field">
-          <select-field
-            v-model="form.stellar.assetType"
-            name="create-stellar-asset-type"
-            :label="'create-asset-form.stellar-asset-type-lbl' | globalize"
-            @blur="touchField('form.stellar.assetType')"
-            :error-message="getFieldErrorMessage(
-              'form.stellar.assetType',
-            )"
-            :disabled="isDisabled"
-          >
-            <option
-              v-for="assetType in STELLAR_ASSET_TYPES"
-              :key="assetType.value"
-              :value="assetType.value"
-            >
-              {{ assetType.labelTranslationId | globalize }}
-            </option>
-          </select-field>
-        </div>
-      </div>
-
-      <!-- eslint-disable max-len -->
-      <div class="app__form-row">
-        <div class="app__form-field">
-          <input-field
-            white-autofill
-            v-model="form.stellar.assetCode"
-            name="create-stellar-asset-code"
-            :label="'create-asset-form.stellar-asset-code-lbl' | globalize"
-            @blur="touchField('form.stellar.assetCode')"
-            :error-message="getFieldErrorMessage('form.stellar.assetCode', {
-              length: getAssetCodeMaxLength(),
-              minLength: CREDIT_ALPHANUM12_MIN_LENGTH
-            })"
-            :disabled="isDisabled || !form.stellar.assetType ||
-              form.stellar.assetType === STELLAR_TYPES.native
-            "
-          />
-        </div>
-      </div>
-      <!-- eslint-enable max-len -->
-    </template>
 
     <h3 class="advanced-step-form__subheading app__form-subheading">
       {{ 'create-asset-form.issuance-subheading' | globalize }}
@@ -313,9 +226,6 @@ import config from '@/config'
 import {
   amountRange,
   requiredIf,
-  minLength,
-  maxLength,
-  alphaNum,
 } from '@validators'
 import { vueRoutes } from '@/vue-router/routes'
 import { mapGetters } from 'vuex'
@@ -326,31 +236,7 @@ const EVENTS = {
   updateIsDisabled: 'update:isDisabled',
 }
 
-const STELLAR_ASSET_TYPES = [
-  {
-    labelTranslationId: 'create-asset-form.credit-alphanum4-stellar-asset-type-lbl',
-    value: 'credit_alphanum4',
-  },
-  {
-    labelTranslationId: 'create-asset-form.credit-alphanum12-stellar-asset-type-lbl',
-    value: 'credit_alphanum12',
-  },
-  {
-    labelTranslationId: 'create-asset-form.native-stellar-asset-type-lbl',
-    value: 'native',
-  },
-]
-
-const STELLAR_TYPES = {
-  creditAlphanum4: 'credit_alphanum4',
-  creditAlphanum12: 'credit_alphanum12',
-  native: 'native',
-}
-
-const CREDIT_ALPHANUM4_MAX_LENGTH = 4
 const CREDIT_ALPHANUM12_MIN_LENGTH = 5
-const CREDIT_ALPHANUM12_MAX_LENGTH = 12
-const NATIVE_XLM_TYPE = 'XLM'
 
 export default {
   name: 'advanced-step-form',
@@ -366,7 +252,6 @@ export default {
       isMaxAmountRestricted: false,
       maxIssuanceAmount: '',
       isPreIssuanceEnabled: false,
-      isStellarIntegrationEnabled: false,
       isUsageRestricted: false,
       assetType: store
         ? String(store.getters[vuexTypes.kvAssetTypeDefault])
@@ -374,19 +259,11 @@ export default {
       preIssuanceAssetSigner: '',
       initialPreissuedAmount: '',
       terms: null,
-      stellar: {
-        withdraw: false,
-        deposit: false,
-        assetType: '',
-        assetCode: '',
-      },
     },
     MIN_AMOUNT: config.MIN_AMOUNT,
     MAX_AMOUNT: config.MAX_AMOUNT,
     DOCUMENT_TYPES,
     vueRoutes,
-    STELLAR_ASSET_TYPES,
-    STELLAR_TYPES,
     CREDIT_ALPHANUM12_MIN_LENGTH,
   }),
 
@@ -418,36 +295,7 @@ export default {
             return this.form.isUsageRestricted
           }),
         },
-        stellar: {
-          assetType: {
-            required: requiredIf(function () {
-              return this.form.isStellarIntegrationEnabled
-            }),
-          },
-          assetCode: {
-            required: requiredIf(function () {
-              return this.form.isStellarIntegrationEnabled
-            }),
-          },
-        },
       },
-    }
-    switch (this.form.stellar.assetType) {
-      case STELLAR_TYPES.creditAlphanum4:
-        validations.form.stellar.assetCode.maxLength = maxLength(
-          CREDIT_ALPHANUM4_MAX_LENGTH
-        )
-        validations.form.stellar.assetCode.alphaNum = alphaNum
-        break
-      case STELLAR_TYPES.creditAlphanum12:
-        validations.form.stellar.assetCode.minLength = minLength(
-          CREDIT_ALPHANUM12_MIN_LENGTH
-        )
-        validations.form.stellar.assetCode.maxLength = maxLength(
-          CREDIT_ALPHANUM12_MAX_LENGTH
-        )
-        validations.form.stellar.assetCode.alphaNum = alphaNum
-        break
     }
     return validations
   },
@@ -459,14 +307,6 @@ export default {
       vuexTypes.kvAssetTypeSecurity,
       vuexTypes.accountId,
     ]),
-  },
-
-  watch: {
-    'form.stellar.assetType' (val) {
-      if (val === STELLAR_TYPES.native) {
-        this.form.stellar.assetCode = NATIVE_XLM_TYPE
-      }
-    },
   },
 
   created () {
@@ -482,8 +322,6 @@ export default {
 
       const isPreIssuanceEnabled =
         this.request.preIssuanceAssetSigner !== config.NULL_ASSET_SIGNER
-
-      const isStellarIntegrationEnabled = Boolean(this.request.stellarAssetCode)
 
       const isUsageRestricted = this.request.assetType &&
         this.request.assetType !== this.kvAssetTypeDefault
@@ -505,13 +343,6 @@ export default {
         terms: this.request.termsKey
           ? new DocumentContainer(this.request.terms)
           : null,
-        isStellarIntegrationEnabled: isStellarIntegrationEnabled,
-        stellar: {
-          withdraw: this.request.stellarWithdraw,
-          deposit: this.request.stellarDeposit,
-          assetType: this.request.stellarAssetType,
-          assetCode: this.request.stellarAssetCode,
-        },
       }
     },
 
@@ -532,16 +363,6 @@ export default {
 
     emitEnabledState () {
       this.$emit(EVENTS.updateIsDisabled, false)
-    },
-
-    getAssetCodeMaxLength () {
-      const assetType = this.form.stellar.assetType
-
-      if (assetType === STELLAR_TYPES.creditAlphanum4) {
-        return CREDIT_ALPHANUM4_MAX_LENGTH
-      } else if (assetType === STELLAR_TYPES.creditAlphanum12) {
-        return CREDIT_ALPHANUM12_MAX_LENGTH
-      }
     },
 
     translateAssetType (value) {